requirements.txt
django
environs
django-jazzmin
django-ckeditor
django-ckeditor-5
django-anymail
django-bootstrap4
crispy-bootstrap4
crispy-bootstrap5
django-recaptcha==3.0.0
django-extensions
shortuuid
django-import-export
stripe
whitenoise
